Fixed some warnings in Assignment 01 and added 02

+++ section-assignment-01/problem01.cpp
@@ -10,7 +10,7 @@ using namespace std;
 void CensorStringTwo(string & text, string remove) {
 
     // Solution from Handouts. January 14, 2008
-    for(int i = 0; i < remove.length(); i++)
+    for(size_t i = 0; i < remove.length(); i++)
     {
         int pos = 0;
         while ((pos = text.find(remove[i], pos)) != string::npos)
